just moved component folder
